requirements.txt
docx==0.2.4
protobuf==6.31.0
Pygments==2.19.1
PyQt6==6.9.0
PyQt6_sip==13.10.0
python-dotenv==1.1.0
QDarkStyle==3.2.3
reportlab==4.4.1
